UEXT and ThingSet code moved to ext folder
