+++ src/ext/README.md
@@ -1,0 +1,3 @@
+# Folder src/ext
+
+This folder contains code for all **ext**ensions (displays etc. connected to UEXT port) and **ext**ernal communication interfaces (UART or CAN).
